third

+++ src/main/java/kr/io/classicgame/service/UserServiceImpl.java
@@ -2,8 +2,6 @@ package kr.io.classicgame.service;
 
 import java.util.Optional;
 
-import javax.transaction.Transactional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -24,12 +22,11 @@ public class UserServiceImpl implements UserService {
 			return null;
 		}
 	}
-	
-	@Transactional
+
 	public boolean insertUser(Usertable user) {
 		// null인 경우 @ExceptionHandler를 통해 예외처리 필요. 닉네임, mail도 검증 필요.
 		boolean result = false;
-		
+
 		if (getUser(user) == null) {
 			userRepo.save(user);
 			result = true;
@@ -37,20 +34,30 @@ public class UserServiceImpl implements UserService {
 		return result;
 	}
 
+	public boolean deleteUser(Usertable user) {
 
-	public void updateUser(Usertable user) {
-		Usertable findUser = userRepo.findById(user.getId()).get();
-		
-		findUser.setName(user.getName());
-		findUser.setPw(user.getPw());
-		findUser.setMail(user.getMail());
-		findUser.setNickname(user.getNickname());
-		
-		userRepo.save(findUser);
+		boolean result = false;
+
+		if (getUser(user) != null) {
+			userRepo.deleteById(user.getId());
+			result = true;
+		}
+		return result;
 	}
 
-	public void deleteUser(Usertable user) {
-		userRepo.deleteById(user.getId());
-	}
+	public boolean updateUser(Usertable user) {
 
+		boolean result = false;
+
+		if (getUser(user) != null) {
+			Usertable findUser = userRepo.findById(user.getId()).get();
+			findUser.setNickname(user.getNickname());
+			findUser.setName(user.getName());
+			findUser.setPw(user.getPw());
+			findUser.setMail(user.getMail());
+			userRepo.save(findUser);
+			result = true;
+		}
+		return result;
+	}
 }
